+++ packages/components/addon/components/hds/disclosure-primitive/index.js
@@ -9,7 +9,7 @@ import { action } from '@ember/object';
 import { schedule } from '@ember/runloop';
 
 export default class HdsDisclosurePrimitiveComponent extends Component {
-  @tracked isOpen; // notice: if in the future we need to add a "@isOpen" prop to control the status from outside (eg to have the DisclosurePrimitive opened on render) just add  "this.args.isOpen" here to initalize the variable
+  @tracked isOpen = this.args.isOpen ?? false;
 
   @action
   onClickToggle() {