+++ packages/components/tests/integration/components/hds/accordion/index-test.js
@@ -8,18 +8,11 @@ import { setupRenderingTest } from 'ember-qunit';
 import { render } from '@ember/test-helpers';
 import { hbs } from 'ember-cli-htmlbars';
 
-module(
-  'Integration | Component | hds/accordion/index',
-  function (hooks) {
-    setupRenderingTest(hooks);
+module('Integration | Component | hds/accordion/index', function (hooks) {
+  setupRenderingTest(hooks);
 
-    test('it should render the component with a CSS class that matches the component name', async function (assert) {
-      await render(
-        hbs`<Hds::Accordion id="test-accordion" />`
-      );
-      assert
-        .dom('#test-accordion')
-        .hasClass('hds-accordion');
-    });
-  }
-);
+  test('it should render the component with a CSS class that matches the component name', async function (assert) {
+    await render(hbs`<Hds::Accordion id="test-accordion" />`);
+    assert.dom('#test-accordion').hasClass('hds-accordion');
+  });
+});